+++ Changelog.md
@@ -6,6 +6,7 @@
 - Enable plotly rendering for jupyter notebooks (#6871)
 - Bug Fix: Prevent assigning inactive graders to a group (#6801)
 - Added new API routes for the `create` and `destroy` actions of SectionsController (#6873)
+- Display error message in real-time when running automated tests, rather than waiting for page refresh (#6878)
 
 ## [v2.4.1]
 - Internal changes only